# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CropPilot AI farmer advisory API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a farmer account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Sign in
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Get the signed-in user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /advice:
    get:
      operationId: listAdvice
      tags: [advice]
      summary: List saved advice reports
      responses:
        "200":
          description: Advice list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdviceReport"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      operationId: createAdvice
      tags: [advice]
      summary: Generate personalized farm advice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdviceInput"
      responses:
        "201":
          description: Advice report saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdviceReport"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /advice/{id}:
    parameters:
      - $ref: "#/components/parameters/Id"
    get:
      operationId: getAdvice
      tags: [advice]
      summary: Get a saved advice report
      responses:
        "200":
          description: Advice details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdviceReport"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /chat/sessions:
    get:
      operationId: listChatSessions
      tags: [chat]
      summary: List chat sessions
      responses:
        "200":
          description: Chat sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatSession"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      operationId: createChatSession
      tags: [chat]
      summary: Start a chat session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatSessionInput"
      responses:
        "201":
          description: New session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatSession"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /chat/sessions/{id}/messages:
    parameters:
      - $ref: "#/components/parameters/Id"
    get:
      operationId: listChatMessages
      tags: [chat]
      summary: Get messages in a session
      responses:
        "200":
          description: Chat messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatMessage"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    post:
      operationId: sendChatMessage
      tags: [chat]
      summary: Send a message to the agronomist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatMessageInput"
      responses:
        "201":
          description: Saved assistant reply
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatMessage"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
...
components:
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    BadRequest:
      description: Invalid input
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Conflict:
      description: Resource conflict
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    RateLimited:
      description: Too many requests
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    ServiceUnavailable:
      description: AI service is unavailable
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
      required: [error]
    User:
      type: object
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        fullName: { type: string }
        phone: { type: ["string", "null"] }
        preferredLanguage: { type: string, enum: [en, hi] }
        regionState: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
      required: [id, email, fullName, phone, preferredLanguage, regionState, createdAt]
    RegisterInput:
      type: object
      properties:
        fullName: { type: string, minLength: 2, maxLength: 120 }
        email: { type: string, format: email, maxLength: 254 }
        password: { type: string, minLength: 10, maxLength: 128 }
        phone: { type: ["string", "null"], maxLength: 32 }
        preferredLanguage: { type: string, enum: [en, hi] }
        regionState: { type: ["string", "null"], maxLength: 100 }
      required: [fullName, email, password, preferredLanguage]
    LoginInput:
      type: object
      properties:
        email: { type: string, format: email, maxLength: 254 }
        password: { type: string, minLength: 1, maxLength: 128 }
      required: [email, password]
...
    AdviceResult:
      type: object
      properties:
        cropRecommendations: { type: array, items: { type: string } }
        fertilizerPlan: { type: array, items: { type: string } }
        irrigationPlan: { type: array, items: { type: string } }
        soilActions: { type: array, items: { type: string } }
        riskNotes: { type: array, items: { type: string } }
        disclaimer: { type: string }
      required: [cropRecommendations, fertilizerPlan, irrigationPlan, soilActions, riskNotes, disclaimer]
    AdviceReport:
      type: object
      properties:
        id: { type: string, format: uuid }
        farmId: { type: ["string", "null"], format: uuid }
        farmName: { type: ["string", "null"] }
        goal: { type: string, enum: [yield, cost, organic, drought] }
        season: { type: string }
        currentCrops: { type: ["string", "null"] }
        specificQuestions: { type: ["string", "null"] }
        result: { $ref: "#/components/schemas/AdviceResult" }
        language: { type: string, enum: [en, hi] }
        createdAt: { type: string, format: date-time }
      required: [id, farmId, farmName, goal, season, currentCrops, specificQuestions, result, language, createdAt]
    AdviceInput:
      type: object
      properties:
        farmId: { type: string, format: uuid }
        goal: { type: string, enum: [yield, cost, organic, drought] }
        season: { type: string, minLength: 1, maxLength: 80 }
        currentCrops: { type: ["string", "null"], maxLength: 1000 }
        specificQuestions: { type: ["string", "null"], maxLength: 2000 }
        language: { type: string, enum: [en, hi] }
      required: [farmId, goal, season, language]
    ChatSession:
      type: object
      properties:
        id: { type: string, format: uuid }
        title: { type: string }
        language: { type: string, enum: [en, hi] }
        updatedAt: { type: string, format: date-time }
        lastMessage: { type: ["string", "null"] }
      required: [id, title, language, updatedAt, lastMessage]
    ChatSessionInput:
      type: object
      properties:
        title: { type: string, maxLength: 120 }
        language: { type: string, enum: [en, hi] }
      required: [language]
    ChatMessage:
      type: object
      properties:
        id: { type: string, format: uuid }
        sessionId: { type: string, format: uuid }
        role: { type: string, enum: [user, assistant] }
        content: { type: string }
        createdAt: { type: string, format: date-time }
      required: [id, sessionId, role, content, createdAt]
    ChatMessageInput:
      type: object
      properties:
        content: { type: string, minLength: 1, maxLength: 4000 }
        useFarmContext: { type: boolean }
      required: [content, useFarmContext]
...