# lib/api-zod/src/generated/types/chatSessionInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * CropPilot AI farmer advisory API
 * OpenAPI spec version: 0.1.0
 */
import type { ChatSessionInputLanguage } from './chatSessionInputLanguage';

export interface ChatSessionInput {
  /** @maxLength 120 */
  title?: string;
  language: ChatSessionInputLanguage;
}