# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * CropPilot AI farmer advisory API
 * OpenAPI spec version: 0.1.0
 */
import type { UserPreferredLanguage } from './userPreferredLanguage';

export interface User {
  id: string;
  email: string;
  fullName: string;
  /** @nullable */
  phone: string | null;
  preferredLanguage: UserPreferredLanguage;
  /** @nullable */
  regionState: string | null;
  createdAt: Date;
}